requirements.txt
pyswip==0.3.2
PyYAML==6.0.2
customtkinter==5.2.2
numpy==2.2.6
pandas==2.3.0
